emai-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
google-cloud-firestore==2.16.0
google-cloud-pubsub==2.23.1
firebase-admin==6.5.0
python-dotenv==1.0.1
